simplified interface folder, added __all__ headers, and moved tacs driver to driver/
