+++ pyfuntofem/driver/funtofem_nlbgs_driver.py
@@ -25,7 +25,7 @@ __all__ = ["FUNtoFEMnlbgs"]
 import numpy as np
 from mpi4py import MPI
 from funtofem import TransferScheme
-from .funtofem_driver import FUNtoFEMDriver
+from ._funtofem_driver import FUNtoFEMDriver
 
 try:
     from .hermes_transfer import HermesTransfer
