--- pyfuntofem/interface/_cart3d/__init__.py
@@ -1,8 +1,0 @@
-# __init__.py of Funtofem cart3d discipline folder
-# imports all cart3d python files to the package
-
-# the packages imported by * in each file are detailed
-# at the top of the file in __all__ = ["Class1", ...]
-
-from .cart3d_interface import *
-from .cart3d_utils import *
