+++ pyfuntofem/interface/_solver_interface.py
@@ -19,6 +19,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+__all__ = ["SolverInterface"]
+
 import numpy as np
