--- pyfuntofem/interface/_test/__init__.py
@@ -1,5 +1,0 @@
-# __init__ file in pyfuntofem/_test discipline
-# Includes Piston theory test interface and TestStructural and TestAero Solvers
-
-from .pistontheory_interface import *
-from .test_solver import *
