+++ pyfuntofem/interface/fun3d_interface.py
@@ -27,7 +27,7 @@ import os
 from fun3d.solvers import Flow, Adjoint
 from fun3d import interface
 from funtofem import TransferScheme
-from pyfuntofem.interface.solver_interface import SolverInterface
+from ._solver_interface import SolverInterface
 
 
 class Fun3dInterface(SolverInterface):
