+++ pyfuntofem/interface/su2_interface.py
@@ -25,7 +25,7 @@ __all__ = ["SU2Interface"]
 import numpy as np
 import shutil
 from funtofem import TransferScheme
-from pyfuntofem.interface.solver_interface import SolverInterface
+from ._solver_interface import SolverInterface
 from mpi4py import MPI
 
 import pysu2
