+++ pyfuntofem/interface/utils/fun3d_client.py
@@ -36,7 +36,7 @@ from fun3d.mda.fsi.fun3d_aero import (
     AdjointProduct,
 )
 from funtofem import TransferScheme
-from pyfuntofem.interface.solver_interface import SolverInterface
+from ._solver_interface import SolverInterface
 
 
 class Fun3dClient(SolverInterface):
