+++ pyfuntofem/model/_base.py
