+++ pyfuntofem/model/body.py
@@ -21,7 +21,7 @@ limitations under the License.
 """
 
 import numpy as np
-from .base import Base
+from ._base import Base
 from mpi4py import MPI
 from funtofem import TransferScheme
