+++ pyfuntofem/optimization/pyopt_optimization.py
@@ -20,6 +20,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
+__all__ = ["PyOptOptimization"]
+
 import numpy as np
